Allow user to adjust profile settings.
The user can now modify the profile settings. The user can modify the
username, email, profile picture, profile tile, and profile body.

Doing this required creating update forms for the user information and
the profile information. "Crispy" forms then takes care of displaying
the fields.

The profile model was modified so that large profile pictures are
resized when they are uploaded to save space.