Make page formatting more consistent.
Cripsy forms allows forms to use bootsrap styling by setting
`CRISPY_TEMPLATE_PACK = 'bootstrap4'`. Also, the pages' contents should
have a margins at the top and bottom so that the contents are not flush
with the edges of the screen. Finally, the database now has profiles for
all the team members.