Add search bar button group.
The button group ensures that we can only search Services, Tasks, or
Users in a single search, in a mutually exclusive way. Also, the this
commit adds an ordering drop down box for allowing the users to order
the results. The ordering happens in the backend and is not yet
implemented.