Get user coordinates when searching.
The search page starts finding the visitor coordinates as soon as the
page loads. The coordinates will be used to find the distance to the
address stated in the posts.

As for the `models.py` file, I changed it so that users can have more
than one post and so that the rich text editor can be left blank when
saving.
can be left blank