Retrieve the user's location once per search.
Users' and visitors' locations are stored while using the search page.
This is better than in the past, where the location had to be retrieved
each time a user submitted a query. It was inefficient and mildly
annoying.

The locations are not stored permanently anywhere, and leaving and
returning to the search page will require retrieving the user's
location again.