Add 'Ads by user' functionality.
Previously there was no way to view all the advertisements posted by the
same person. This commit changes that by adding a link in the user's
profile and advertisments that leads visitors to the rest of the user's
posts.

The files `header.html`, `search/urls.py`, and `search/views.py` had
minor changes that can be ignored.

The files `post_detail.html` and `profile_static.html` were cleaned up
and modified to show a link to the rest of the user's advertisements.

The `user/urls.py` specifies that the username is specified in the URL.
The username is used by the `PostListView` function in `user/views.py`
to display all the advertisments whose author's name matches the
username. The advertisments posted by the author are displayed through
the `posts_by_user.html` template.