Add search page pagination.
There was no limit to the number of results that would be shown on the
search page, which could have lead to long loading times if the user
received hundreds of results. Adding pagination fixes that problem. Only
a limited number of results will be displayed per page. The user can go
to next or previous pages to view more results.