Fix orphaned profile picture bug.
Changing the user profile picture resulted in the previous profile
picture not being deleted. Keeping track of the previous profile picture
name solved the issue because the name can be used to delete the
previous profile picture before adding the new profile picture.

There is only one case whenever the previous profile picture should not
be deleted, and that is when the previous profile picture is
`default.png`, since other users may be using it.

The `Profile` model was changed, so I made and ran the needed
migrations.