Lockout user after 3 failed logins.
The `django-axes` library facilitates locking out users. The custom
`login` function in `home/views.py` passes the request parameter to the
`authenticate` function from `django.contrib.auth`. This is needed by
`django-axes`.

The `nifti/settings.py` specifies the number of attempts before lockout,
the duration of the lockout, and the template to show during lockout.
The duration of the lockout is 1 hour, and the template shown is
`home/lockout.html`.