Use one text field to add and remove tags.
Previously, there was a text field and button for adding tags and a
separte text field and button for removing tags. This commit removes one
of the text fields because only one text field is needed. If the add
button is clicked, the tag in the text field is added, and if the remove
button is clicked, the tag in the text field is removed.

Also, since it is inconvenient to have to write the name of a tag you
are trying to remove, I added a function that populates the text field
with the name of the tag whenever the tag is clicked.